Migrate file, public-files, and public-jobs to typed route contracts
