+++ apps/hyperlocalise-web/src/api/routes/file/file.route.test.ts
@@ -1,12 +1,12 @@
 import "dotenv/config";
 
-import { testClient } from "hono/testing";
-import { afterEach, describe, expect, it, vi } from "vite-plus/test";
+import { afterEach, beforeAll, describe, expect, it, vi } from "vite-plus/test";
 
 import { createApp } from "@/api/app";
-import type { FileStorageAdapter, PutStoredObjectInput } from "@/lib/file-storage";
+import { db } from "@/lib/database";
 import { createStoredFile } from "@/lib/file-storage/records";
 import { createProjectTestFixture } from "../project/project.fixture";
+import { createMemoryFileStorageAdapter, ensureStoredFilesTestSchema } from "./file.fixture";
 
 const { resolveApiAuthContextFromSessionMock } = vi.hoisted(() => ({
   resolveApiAuthContextFromSessionMock: vi.fn(() => globalThis.__testApiAuthContext ?? null),
@@ -16,70 +16,14 @@ vi.mock("@/api/auth/workos-session", () => ({
   resolveApiAuthContextFromSession: resolveApiAuthContextFromSessionMock,
 }));
 
-function createMemoryFileStorageAdapter(): FileStorageAdapter {
-  const store = new Map<string, { buffer: Buffer; contentType: string }>();
-
-  return {
-    provider: "vercel_blob",
-    async put(input: PutStoredObjectInput) {
-      let buffer: Buffer;
-      if (input.body instanceof ReadableStream) {
-        const chunks: Uint8Array[] = [];
-        const reader = input.body.getReader();
-        while (true) {
-          const { done, value } = await reader.read();
-          if (done) break;
-          if (value) chunks.push(value);
-        }
-        buffer = Buffer.concat(chunks.map((c) => Buffer.from(c)));
-      } else if (Buffer.isBuffer(input.body)) {
-        buffer = input.body;
-      } else if (input.body instanceof ArrayBuffer) {
-        buffer = Buffer.from(input.body);
-      } else if (input.body instanceof Uint8Array) {
-        buffer = Buffer.from(input.body.buffer, input.body.byteOffset, input.body.byteLength);
-      } else {
-        buffer = Buffer.from(await input.body.arrayBuffer());
-      }
-
-      store.set(input.key, { buffer, contentType: input.contentType });
-
-      return {
-        provider: "vercel_blob",
-        key: input.key,
-        url: `https://blob.example/${input.key}`,
-        downloadUrl: `https://blob.example/${input.key}?download=1`,
-        contentType: input.contentType,
-        etag: "test-etag",
-      };
-    },
-    async get(input) {
-      const entry = store.get(input.keyOrUrl);
-      if (!entry) return null;
-      return {
-        body: new ReadableStream({
-          start(controller) {
-            controller.enqueue(new Uint8Array(entry.buffer));
-            controller.close();
-          },
-        }),
-        contentType: entry.contentType,
-        etag: "test-etag",
-      };
-    },
-    async delete(input) {
-      store.delete(input.keyOrUrl);
-    },
-    async getSignedUrl(input) {
-      return `https://blob.example/${input.keyOrUrl}?signed=1`;
-    },
-  };
-}
-
 const fileStorageAdapter = createMemoryFileStorageAdapter();
 const app = createApp({ fileStorageAdapter });
-const client = testClient(app);
-const { authHeadersFor, cleanup, createWorkosIdentity } = createProjectTestFixture(client);
+const { authHeadersFor, cleanup, createWorkosIdentity } = createProjectTestFixture();
+
+beforeAll(async () => {
+  await db.$client.query("select 1");
+  await ensureStoredFilesTestSchema();
+});
 
 afterEach(async () => {
   vi.clearAllMocks();
