+++ apps/hyperlocalise-web/src/api/routes/file/file.shared.ts
@@ -1,0 +1,3 @@
+export function fileNotFoundResponse(c: { json(body: { error: string }, status: 404): Response }) {
+  return c.json({ error: "not_found" }, 404);
+}
